# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.3.0",
    "description": "Go-native Rebecca panel API for dashboard, bot, and third-party client developers.\n\nAuthentication: the dashboard uses an opaque server-side session stored in the HttpOnly `rebecca_session` cookie. Session permissions are loaded from the database on every request, so permission changes take effect without logging out. Bots and external integrations should use personal `rk_...` API keys as `Authorization: Bearer <api_key>`. Legacy JWT creation at `POST /admin/token` and `POST /api/admin/token` remains available only for older Marzban-compatible clients.\n\nDo not automate username/password login for bots. API keys can be revoked independently and inherit the owner admin permissions, service scope, and limits.\n\nMutation behavior: user, service, node, config, and host mutations are transactional where applicable. Runtime changes are synchronized to nodes through gRPC and recorded for recovery."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/maintenance/update": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Schedule panel update",
        "description": "Schedule panel update on binary installations.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: System stats require an authenticated admin. Maintenance actions require sudo/full_access and binary runtime.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nThe response includes an operation snapshot. Poll /api/maintenance/status until the operation enters restarting, then wait for /api/maintenance/info to respond again and refresh the dashboard.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "description": "Operation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "action": {
                          "type": "string"
                        },
                        "phase": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "progress": {
                          "type": "integer",
                          "nullable": true,
                          "minimum": 0,
                          "maximum": 100
                        },
                        "running": {
                          "type": "boolean"
                        },
                        "restarting": {
                          "type": "boolean"
                        },
                        "needs_reload": {
                          "type": "boolean"
                        },
                        "error": {
                          "type": "string"
                        },
                        "logs": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "started_at": {
                          "type": "integer"
                        },
                        "updated_at": {
                          "type": "integer"
                        },
                        "finished_at": {
                          "type": "integer",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "202": {
            "$ref": "#/components/responses/JSON"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postMaintenanceUpdate"
      }
    },
    "/api/maintenance/restart": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Schedule panel restart",
        "description": "Schedule panel restart on binary installations.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: System stats require an authenticated admin. Maintenance actions require sudo/full_access and binary runtime.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nThe response includes an operation snapshot. Poll /api/maintenance/status until the operation enters restarting, then wait for /api/maintenance/info to respond again and refresh the dashboard.",
        "responses": {
          "200": {
            "description": "Operation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "action": {
                          "type": "string"
                        },
                        "phase": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "progress": {
                          "type": "integer",
                          "nullable": true,
                          "minimum": 0,
                          "maximum": 100
                        },
                        "running": {
                          "type": "boolean"
                        },
                        "restarting": {
                          "type": "boolean"
                        },
                        "needs_reload": {
                          "type": "boolean"
                        },
                        "error": {
                          "type": "string"
                        },
                        "logs": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "started_at": {
                          "type": "integer"
                        },
                        "updated_at": {
                          "type": "integer"
                        },
                        "finished_at": {
                          "type": "integer",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "202": {
            "$ref": "#/components/responses/JSON"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postMaintenanceRestart"
      }
    },
    "/api/maintenance/soft-reload": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Schedule panel soft reload",
        "description": "Schedule a panel-only reload/restart without touching node Xray runtimes.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: System stats require an authenticated admin. Maintenance actions require sudo/full_access and binary runtime.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nThe response includes an operation snapshot. Poll /api/maintenance/status until the operation enters restarting, then wait for /api/maintenance/info to respond again and refresh the dashboard.",
        "responses": {
          "200": {
            "description": "Operation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "action": {
                          "type": "string"
                        },
                        "phase": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "progress": {
                          "type": "integer",
                          "nullable": true,
                          "minimum": 0,
                          "maximum": 100
                        },
                        "running": {
                          "type": "boolean"
                        },
                        "restarting": {
                          "type": "boolean"
                        },
                        "needs_reload": {
                          "type": "boolean"
                        },
                        "error": {
                          "type": "string"
                        },
                        "logs": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "started_at": {
                          "type": "integer"
                        },
                        "updated_at": {
                          "type": "integer"
                        },
                        "finished_at": {
                          "type": "integer",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "202": {
            "$ref": "#/components/responses/JSON"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postMaintenanceSoftReload"
      }
    },
...
    "/api/maintenance/status": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Get latest panel maintenance operation status",
        "description": "Returns the latest host-level update, restart, or soft-reload operation tracked by the Go API. Dashboard clients poll this endpoint while the maintenance dialog is open to show command output, parsed download percentage, current phase, and whether the panel is entering restart. The state is intentionally in-memory and resets after the Rebecca process comes back.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Latest maintenance operation status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "action": {
                      "type": "string"
                    },
                    "phase": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "progress": {
                      "type": "integer",
                      "nullable": true,
                      "minimum": 0,
                      "maximum": 100
                    },
                    "running": {
                      "type": "boolean"
                    },
                    "restarting": {
                      "type": "boolean"
                    },
                    "needs_reload": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    },
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "started_at": {
                      "type": "integer"
                    },
                    "updated_at": {
                      "type": "integer"
                    },
                    "finished_at": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "getMaintenanceStatus"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Rebecca admin API key or legacy JWT",
        "description": "Use a personal rk_ API key for bots and third-party integrations. Legacy JWTs from /admin/token remain accepted for older Marzban-compatible clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rebecca_session",
        "description": "Opaque HttpOnly dashboard session cookie issued by POST /api/auth/login. It is managed by the browser and is not readable by dashboard JavaScript."
      }
    },
...
    "requestBodies": {
      "JSON": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Object"
            },
            "example": {
              "name": "example",
              "status": "active",
              "settings": {}
            }
          }
        },
        "description": "Generic JSON request body. See each operation description for required business fields and side effects."
      },
...
    },
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible Rebecca API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            },
            "example": {
              "id": 1,
              "status": "ok",
              "detail": "Operation completed"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response. Most handlers return `{ \"detail\": \"...\" }`; some compatibility routes may return a plain text body when handled by net/http helpers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "additionalProperties": true,
        "example": {
          "name": "example",
          "status": "active",
          "settings": {}
        }
      },
      "AnyValue": {
        "nullable": true,
        "example": {
          "id": 1,
          "status": "ok"
        },
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}